requirements.txt
Django>=4.2,<5.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-tables2==2.5.3
django-tables2-bootstrap4==2.0.0b5
django-bootstrap4==23.1
django-bootstrap-datepicker-plus==3.0.5
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-debug-toolbar==4.1.0
django-scheduler
django-widget-tweaks
django-hijack
django-jquery
django-admin-autocomplete-filter==0.7.1
django-currentuser==0.7.0
psycopg2>=2.7,<3.0
PyYAML
yamlns
rq
rq-scheduler
django-rq
django-anymail
sendgrid
pymongo==3.10.1
django-rosetta
sentry-sdk
connexion
python-jose
djangorestframework
django-cors-headers
uwsgi
djangorestframework-simplejwt
